Move participant creation logic into handler from BandUtil
