+++ lib/features/modeling/ChoreoModeling.js
@@ -149,7 +149,7 @@ ChoreoModeling.prototype.unlinkCallChoreoParticipants = function(element) {
  */
 ChoreoModeling.prototype.createParticipant = function() {
   const participantContext = {};
-  this._commandStack.execute('participant.create',participantContext);
+  this._commandStack.execute('participant.create', participantContext);
   return participantContext.created;
 };
