+++ lib/features/modeling/cmd/CreateParticipantBandHandler.js
@@ -37,10 +37,9 @@ import { removeAndGetIndex } from '../../../util/MessageUtil';
  * @param {ElementFactory} elementFactory
  * @param {Moddle} moddle
  * @param {ChoreoModeling} modeling
- * @param {CommandStack} commandStack
  */
 export default function CreateParticipantBandHandler(
-    injector, canvas, bpmnFactory, elementFactory, moddle, modeling, commandStack
+    injector, canvas, bpmnFactory, elementFactory, moddle, modeling
 ) {
   this._injector = injector;
   this._canvas = canvas;
@@ -48,7 +47,6 @@ export default function CreateParticipantBandHandler(
   this._elementFactory = elementFactory;
   this._model = moddle;
   this._modeling = modeling;
-  this._commandStack = commandStack;
 }
 
 CreateParticipantBandHandler.$inject = [
