+++ lib/util/BandUtil.js
@@ -44,36 +44,6 @@ export function updateParticipantBandKinds(activity, diBands) {
 }
 
 /**
- * Creates a new participant and adds it to the choreography.
- * This is not a command, for the command version call the function in ChoreoModeling, or directly via the CommandStack
- * with participant.create
- */
-export function createParticipant(injector, attrs) {
-  let participant = injector.get('bpmnFactory').create('bpmn:Participant', attrs);
-
-  let choreo = injector.get('canvas').getRootElement();
-  if (choreo.businessObject.participants) {
-    choreo.businessObject.participants.push(participant);
-  } else {
-    choreo.businessObject.participants = [ participant ];
-  }
-  participant.name = 'New Participant ' + countNewParticipants(choreo.businessObject.participants);
-  participant.$parent = choreo;
-  return participant;
-}
-
-/**
- *
- * @param {Array} participants
- */
-function countNewParticipants(participants) {
-  if (participants) {
-    return participants.filter(p => p.name).filter(p => p.name.indexOf('New Participant') === 0).length + 1;
-  }
-  return 1;
-}
-
-/**
  * Resets and recalculates the bounds and other properties of the participant bands
  * of an activity *on a DI level*, not on shapes.
  *
